Standardized on passing in units in kg/m/s space.
No conversions happen inside the sandbox. Instead, scale is stored
for making the conversions easier on users and for debug drawing.
Otherwise, it's up to the user to decide how they want to scale things.
The scale factor just makes it easier to get started, but stays
out of the way in bigger projects.
SandboxApp updated to reflect changes in API.