Separated scaling from the sandbox.
Makes it easy to write your own scaling functions if desired; there
is nothing to replace/work around.
Scale class makes it easy to convert things and is a good place to
start when considering the relationship between your screen space
graphics and physics space bodies. I recommend using it, but you're
not bound to it like you were when it lived inside of Sandbox.
Updated drawing sample to work with new scaling location.